requirements.txt
torch==1.8.1
torchvision==0.9.1
pytorch-lightning==1.3.7
ftfy
regex
tqdm
omegaconf
ffmpeg-python
scipy
scikit-image
scikit-video
einops
gdown
transformers
requests